feat: use new input component everywhere
