+++ src/app/Components/ArtistAutosuggest/ArtistAutosuggest.tsx
@@ -1,6 +1,5 @@
-import { Box } from "@artsy/palette-mobile"
+import { Box, Input2 } from "@artsy/palette-mobile"
 import SearchIcon from "app/Components/Icons/SearchIcon"
-import { Input } from "app/Components/Input"
 import { SearchContext, useSearchProviderValues } from "app/Scenes/Search/SearchContext"
 import { ArtworkDetailsFormModel } from "app/Scenes/SellWithArtsy/SubmitArtwork/ArtworkDetails/validation"
 import { useFormikContext } from "formik"
@@ -50,7 +49,7 @@ export const ArtistAutosuggest: React.FC<ArtistAutosuggestProps> = ({
 
   return (
     <SearchContext.Provider value={searchProviderValues}>
-      <Input
+      <Input2
         title={title || undefined}
         placeholder={placeholder}
         icon={<SearchIcon width={18} height={18} />}
