+++ src/app/Components/ArtworkFilter/Filters/CustomSizeInputs.tsx
@@ -1,8 +1,6 @@
-import { Box, Flex, Join, Spacer, Text, useColor } from "@artsy/palette-mobile"
-import { Input } from "app/Components/Input"
+import { Box, Flex, Input2, Join, Spacer, Text } from "@artsy/palette-mobile"
 import { Metric } from "app/Scenes/Search/UserPrefsModel"
 import React, { useState } from "react"
-import { TextStyle } from "react-native"
 import { Numeric, Range } from "./helpers"
 
 export interface CustomSizeInputsProps {
@@ -33,11 +31,9 @@ const getValue = (value: Numeric) => {
 export const CustomSizeInputs: React.FC<CustomSizeInputsProps> = ({
   label,
   range,
-  active,
   onChange,
   selectedMetric,
 }) => {
-  const color = useColor()
   const [localRange, setLocalRange] = useState(range)
 
   const handleInputChange = (field: keyof Range) => (text: string) => {
@@ -52,10 +48,6 @@ export const CustomSizeInputs: React.FC<CustomSizeInputsProps> = ({
     onChange({ ...range, [field]: value })
   }
 
-  const inputTextStyle: TextStyle = {
-    color: active ? color("black100") : color("black60"),
-  }
-
   return (
     <Box>
       <Text variant="xs" caps mb={0.5}>
@@ -64,27 +56,23 @@ export const CustomSizeInputs: React.FC<CustomSizeInputsProps> = ({
       <Flex flexDirection="row">
         <Join separator={<Spacer x={2} />}>
           <Flex flex={1}>
-            <Input
-              description="Min"
-              descriptionColor="black60"
+            <Input2
+              title="Min"
               keyboardType="number-pad"
               onChangeText={handleInputChange("min")}
               fixedRightPlaceholder={selectedMetric}
               accessibilityLabel={`Minimum ${label} Input`}
               value={getValue(localRange.min)}
-              inputTextStyle={inputTextStyle}
             />
           </Flex>
           <Flex flex={1}>
-            <Input
-              description="Max"
-              descriptionColor="black60"
+            <Input2
+              title="Max"
               keyboardType="number-pad"
               onChangeText={handleInputChange("max")}
               fixedRightPlaceholder={selectedMetric}
               accessibilityLabel={`Maximum ${label} Input`}
               value={getValue(localRange.max)}
-              inputTextStyle={inputTextStyle}
             />
           </Flex>
         </Join>
