+++ src/app/Components/LocationAutocomplete.tsx
@@ -1,21 +1,21 @@
-import { MapPinIcon, Flex, Text, Touchable } from "@artsy/palette-mobile"
-import { Input, InputProps } from "app/Components/Input"
+import { Flex, Input2, Input2Props, MapPinIcon, Text, Touchable } from "@artsy/palette-mobile"
+import { Input } from "app/Components/Input"
 import {
-  getLocationDetails,
-  getLocationPredictions,
   LocationWithDetails,
   SimpleLocation,
+  getLocationDetails,
+  getLocationPredictions,
 } from "app/utils/googleMaps"
 import { useScreenDimensions } from "app/utils/hooks"
 import React, { useEffect, useRef, useState } from "react"
 import { TouchableWithoutFeedback } from "react-native"
 
-interface LocationAutocompleteProps extends Omit<InputProps, "onChange"> {
+interface LocationAutocompleteProps extends Omit<Input2Props, "onChange"> {
   floating?: boolean
   initialLocation?: LocationWithDetails | null
   displayLocation?: string
   allowCustomLocation?: boolean
-  inputRef?: React.RefObject<Input>
+  inputRef?: React.RefObject<Input2>
   showError?: boolean
   onChange: (l: LocationWithDetails) => void
   FooterComponent?: () => JSX.Element
@@ -115,7 +115,7 @@ export const LocationAutocomplete: React.FC<LocationAutocompleteProps> = ({
 
   return (
     <Flex>
-      <Input
+      <Input2
         {...restProps}
         ref={ref}
         testID="autocomplete-location-input"
