+++ src/app/Scenes/MyAccount/MyAccountEditPassword.tsx
@@ -1,9 +1,8 @@
-import { Flex, Separator } from "@artsy/palette-mobile"
-import { Input } from "app/Components/Input"
+import { Flex, Input2, Separator } from "@artsy/palette-mobile"
 import { Stack } from "app/Components/Stack"
 import { getCurrentEmissionState, GlobalStore, unsafe__getEnvironment } from "app/store/GlobalStore"
 import React, { useEffect, useState } from "react"
-import { Alert } from "react-native"
+import { Alert, Text } from "react-native"
 import {
   MyAccountFieldEditScreen,
   MyAccountFieldEditScreenProps,
@@ -95,11 +94,10 @@ export const MyAccountEditPassword: React.FC<{}> = ({}) => {
       contentContainerStyle={{ paddingHorizontal: 0 }}
     >
       <Flex mx={2}>
-        <Input
+        <Input2
           autoComplete="password"
           autoFocus
           onChangeText={setCurrentPassword}
-          placeholder="Current password"
           secureTextEntry
           enableClearButton
           title="Current password"
@@ -109,19 +107,21 @@ export const MyAccountEditPassword: React.FC<{}> = ({}) => {
       </Flex>
       <Separator mb={2} mt={4} />
       <Stack mx={2}>
-        <Input
-          description="Must include at least one uppercase letter, one lowercase letter, and one number."
+        <Text>
+          Password must include at least one uppercase letter, one lowercase letter, and one number.
+        </Text>
+        <Input2
+          // TODO: Look into this with design
+
           onChangeText={setNewPassword}
-          placeholder="New password"
           secureTextEntry
           enableClearButton
           title="New password"
           value={newPassword}
           error={receivedErrorNew}
         />
-        <Input
+        <Input2
           onChangeText={setPasswordConfirmation}
-          placeholder="Confirm new password"
           secureTextEntry
           enableClearButton
           title="Confirm new password"
