+++ src/app/Scenes/MyCollection/Screens/ArtworkForm/Components/ArtworkAutosuggest.tsx
@@ -1,6 +1,5 @@
-import { Flex, Text, Touchable } from "@artsy/palette-mobile"
+import { Flex, Input2, Text, Touchable } from "@artsy/palette-mobile"
 import SearchIcon from "app/Components/Icons/SearchIcon"
-import { Input } from "app/Components/Input"
 import { useArtworkForm } from "app/Scenes/MyCollection/Screens/ArtworkForm/Form/useArtworkForm"
 import React, { useState } from "react"
 import { ArtworkAutosuggestResultsQueryRenderer } from "./ArtworkAutosuggestResults"
@@ -27,7 +26,7 @@ export const ArtworkAutosuggest: React.FC<ArtworkAutosuggestProps> = ({
 
   return (
     <>
-      <Input
+      <Input2
         placeholder="Search artworks"
         icon={<SearchIcon width={18} height={18} />}
         onChangeText={(value) => {
