+++ src/app/Scenes/MyProfile/MyProfilePaymentNewCreditCard.tsx
@@ -1,11 +1,12 @@
+import { Input2 } from "@artsy/palette-mobile"
 import { MyProfilePaymentNewCreditCardSaveCardMutation } from "__generated__/MyProfilePaymentNewCreditCardSaveCardMutation.graphql"
 import { CountrySelect } from "app/Components/CountrySelect"
-import { Input, InputTitle } from "app/Components/Input"
+import { InputTitle } from "app/Components/Input"
 import { Select } from "app/Components/Select/SelectV2"
 import { Stack } from "app/Components/Stack"
 import { MyAccountFieldEditScreen } from "app/Scenes/MyAccount/Components/MyAccountFieldEditScreen"
 import { getRelayEnvironment } from "app/system/relay/defaultEnvironment"
-import { Action, action, computed, Computed, useLocalStore } from "easy-peasy"
+import { Action, Computed, action, computed, useLocalStore } from "easy-peasy"
 import React, { useEffect, useRef } from "react"
 import { LiteCreditCardInput } from "react-native-credit-card-input"
 import { commitMutation, graphql } from "react-relay"
@@ -84,11 +85,11 @@ export const MyProfilePaymentNewCreditCard: React.FC<{}> = ({}) => {
 
   const paymentInfoRef = useRef<any>(null)
 
-  const addressLine1Ref = useRef<Input>(null)
-  const addressLine2Ref = useRef<Input>(null)
-  const cityRef = useRef<Input>(null)
-  const postalCodeRef = useRef<Input>(null)
-  const stateRef = useRef<Input>(null)
+  const addressLine1Ref = useRef<Input2>(null)
+  const addressLine2Ref = useRef<Input2>(null)
+  const cityRef = useRef<Input2>(null)
+  const postalCodeRef = useRef<Input2>(null)
+  const stateRef = useRef<Input2>(null)
   const countryRef = useRef<Select<any>>(null)
 
   // focus top field on mount
@@ -159,14 +160,14 @@ export const MyProfilePaymentNewCreditCard: React.FC<{}> = ({}) => {
           />
         </>
 
-        <Input
+        <Input2
           title="Name on card"
           placeholder="Full name"
           onChangeText={actions.fields.fullName.setValue}
           returnKeyType="next"
           onSubmitEditing={() => addressLine1Ref.current?.focus()}
         />
-        <Input
+        <Input2
           ref={addressLine1Ref}
           title="Address line 1"
           placeholder="Add street address"
@@ -174,7 +175,7 @@ export const MyProfilePaymentNewCreditCard: React.FC<{}> = ({}) => {
           returnKeyType="next"
           onSubmitEditing={() => addressLine2Ref.current?.focus()}
         />
-        <Input
+        <Input2
           ref={addressLine2Ref}
           title="Address line 2"
           optional
@@ -183,26 +184,23 @@ export const MyProfilePaymentNewCreditCard: React.FC<{}> = ({}) => {
           returnKeyType="next"
           onSubmitEditing={() => cityRef.current?.focus()}
         />
-        <Input
+        <Input2
           ref={cityRef}
           title="City"
-          placeholder="Add city"
           onChangeText={actions.fields.city.setValue}
           returnKeyType="next"
           onSubmitEditing={() => postalCodeRef.current?.focus()}
         />
-        <Input
+        <Input2
           ref={postalCodeRef}
           title="Postal Code"
-          placeholder="Add postal code"
           onChangeText={actions.fields.postCode.setValue}
           returnKeyType="next"
           onSubmitEditing={() => stateRef.current?.focus()}
         />
-        <Input
+        <Input2
           ref={stateRef}
           title="State, province, or region"
-          placeholder="Add state, province, or region"
           onChangeText={actions.fields.state.setValue}
           onSubmitEditing={() => {
             stateRef.current?.blur()
