+++ src/app/Scenes/Onboarding/ForgotPassword.tsx
@@ -1,14 +1,10 @@
-import { Text, Spacer, Flex, useColor, Button } from "@artsy/palette-mobile"
+import { Input2 } from "@artsy/palette-mobile"
 import { StackScreenProps } from "@react-navigation/stack"
-import { Input } from "app/Components/Input"
 import { GlobalStore } from "app/store/GlobalStore"
-import { BackButton } from "app/system/navigation/BackButton"
-import { ArtsyKeyboardAvoidingView } from "app/utils/ArtsyKeyboardAvoidingView"
-import { useScreenDimensions } from "app/utils/hooks"
-import { FormikProvider, useFormik, useFormikContext } from "formik"
+import { FormikProvider, useFormik } from "formik"
 import React, { useRef, useState } from "react"
-import { ScrollView } from "react-native"
 import * as Yup from "yup"
+import { ForgotPasswordForm } from "./ForgotPasswordForm"
 import { OnboardingNavigationStack } from "./Onboarding"
 
 export const forgotPasswordSchema = Yup.object().shape({
@@ -21,109 +17,7 @@ export interface ForgotPasswordValuesSchema {
 export type ForgotPasswordProps = StackScreenProps<OnboardingNavigationStack, "OnboardingLogin">
 export interface ForgotPasswordFormProps extends ForgotPasswordProps {
   requestedPasswordReset: boolean
-  inputRef?: React.Ref<Input>
-}
-
-export const ForgotPasswordForm: React.FC<ForgotPasswordFormProps> = ({
-  navigation,
-  requestedPasswordReset,
-  inputRef,
-}) => {
-  const { values, handleSubmit, handleChange, validateForm, isValid, dirty, isSubmitting } =
-    useFormikContext<ForgotPasswordValuesSchema>()
-  const color = useColor()
-
-  return (
-    <Flex flex={1} backgroundColor="white" flexGrow={1} pb={1}>
-      <ArtsyKeyboardAvoidingView>
-        <ScrollView
-          contentContainerStyle={{ paddingTop: useScreenDimensions().safeAreaInsets.top }}
-          keyboardShouldPersistTaps="always"
-        >
-          <Flex flex={1} px={2} pt={6} justifyContent="flex-start">
-            <Text variant="lg-display">Forgot Password?</Text>
-            <Text pt={0.5} color="black100" variant="xs">
-              Please enter the email address associated with your Artsy account to receive a reset
-              link.
-            </Text>
-            <Spacer y={2} />
-
-            {!!requestedPasswordReset ? (
-              <Text color="blue100">Password reset link sent. Please check your email.</Text>
-            ) : (
-              <Input
-                ref={inputRef}
-                autoCapitalize="none"
-                autoComplete="email"
-                enableClearButton
-                keyboardType="email-address"
-                onChangeText={(text) => {
-                  handleChange("email")(text.trim())
-                }}
-                onSubmitEditing={() => {
-                  if (dirty) {
-                    handleSubmit()
-                  }
-                }}
-                onBlur={() => {
-                  validateForm()
-                }}
-                blurOnSubmit={false} // This is needed to avoid UI jump when the user submits
-                placeholder="Email address"
-                placeholderTextColor={color("black30")}
-                value={values.email}
-                returnKeyType="done"
-                spellCheck={false}
-                autoCorrect={false}
-                textContentType="emailAddress"
-                testID="email-address"
-              />
-            )}
-          </Flex>
-        </ScrollView>
-        <BackButton onPress={() => navigation.goBack()} />
-        <Flex px={2} paddingBottom={2}>
-          {!!requestedPasswordReset ? (
-            <>
-              <Button
-                variant="fillDark"
-                onPress={() => navigation.goBack()}
-                block
-                haptic="impactMedium"
-                testID="returnToLoginButton"
-              >
-                Return to login
-              </Button>
-              <Spacer y={1} />
-              <Button
-                onPress={handleSubmit}
-                block
-                haptic="impactMedium"
-                disabled={!isValid || !dirty}
-                loading={isSubmitting}
-                testID="resetButton"
-                variant="outline"
-              >
-                Send Again
-              </Button>
-            </>
-          ) : (
-            <Button
-              onPress={handleSubmit}
-              block
-              variant="fillDark"
-              haptic="impactMedium"
-              disabled={!isValid || !dirty}
-              loading={isSubmitting}
-              testID="resetButton"
-            >
-              Send Reset Link
-            </Button>
-          )}
-        </Flex>
-      </ArtsyKeyboardAvoidingView>
-    </Flex>
-  )
+  inputRef?: React.Ref<Input2>
 }
 
 const initialValues: ForgotPasswordValuesSchema = { email: "" }
@@ -131,7 +25,7 @@ const initialValues: ForgotPasswordValuesSchema = { email: "" }
 export const ForgotPassword: React.FC<ForgotPasswordProps> = ({ navigation, route }) => {
   const [requestedPasswordReset, setRequestedPasswordReset] = useState(false)
 
-  const inputRef = useRef<Input>(null)
+  const inputRef = useRef<Input2>(null)
 
   const formik = useFormik<ForgotPasswordValuesSchema>({
     enableReinitialize: true,
