+++ src/app/Scenes/Onboarding/OnboardingCreateAccount/OnboardingCreateAccountEmail.tsx
@@ -1,6 +1,5 @@
-import { useColor } from "@artsy/palette-mobile"
+import { Input2, useColor } from "@artsy/palette-mobile"
 import { StackScreenProps } from "@react-navigation/stack"
-import { Input } from "app/Components/Input"
 import { useFormikContext } from "formik"
 import { Platform } from "react-native"
 import {
@@ -29,7 +28,7 @@ export const OnboardingCreateAccountEmail: React.FC<OnboardingCreateAccountEmail
       onBackButtonPress={route.params.navigateToWelcomeScreen}
       title="Sign Up with Email"
     >
-      <Input
+      <Input2
         autoCapitalize="none"
         autoComplete="email"
         enableClearButton
