+++ src/app/Scenes/SellWithArtsy/ConsignmentInquiry/ConsignmentInquiryForm.tsx
@@ -1,6 +1,5 @@
-import { Spacer, Box, Text, LinkText, Button } from "@artsy/palette-mobile"
+import { Box, Button, Input2, LinkText, Spacer, Text } from "@artsy/palette-mobile"
 import { useNavigation } from "@react-navigation/native"
-import { Input } from "app/Components/Input"
 import { PhoneInput } from "app/Components/Input/PhoneInput"
 import { navigate } from "app/system/navigation/navigate"
 import { useScreenDimensions } from "app/utils/hooks"
@@ -36,10 +35,10 @@ export const ConsignmentInquiryForm: React.FC<{
   }, [canPopScreen, dirty, navigation])
 
   const scrollViewRef = useRef<ScrollView>(null)
-  const nameInputRef = useRef<Input>(null)
-  const emailInputRef = useRef<Input>(null)
-  const phoneInputRef = useRef<Input>(null)
-  const messageInputRef = useRef<Input>(null)
+  const nameInputRef = useRef<Input2>(null)
+  const emailInputRef = useRef<Input2>(null)
+  const phoneInputRef = useRef<Input2>(null)
+  const messageInputRef = useRef<Input2>(null)
 
   const jumpToNextField = (currentField: "name" | "email" | "phone" | "message") => {
     switch (currentField) {
@@ -84,7 +83,7 @@ export const ConsignmentInquiryForm: React.FC<{
           <Text variant="lg-display" mb={2}>
             {!!recipientName ? `Contact ${recipientName}` : "Contact a specialist"}
           </Text>
-          <Input
+          <Input2
             accessibilityLabel="Name"
             autoCapitalize="words"
             autoCorrect={false}
@@ -103,7 +102,7 @@ export const ConsignmentInquiryForm: React.FC<{
             value={values.name}
           />
           <Spacer y={2} />
-          <Input
+          <Input2
             accessibilityLabel="Email address"
             autoCapitalize="none"
             autoCorrect={false}
@@ -139,7 +138,7 @@ export const ConsignmentInquiryForm: React.FC<{
             value={values.phoneNumber}
           />
           <Spacer y={4} />
-          <Input
+          <Input2
             accessibilityLabel="Message to the Artsy Specialist"
             blurOnSubmit={false}
             error={errors.message}
