+++ src/app/system/devTools/DevMenu/Components/NavigateTo.tsx
@@ -1,4 +1,4 @@
-import { Button, Flex, Input, Spacer } from "@artsy/palette-mobile"
+import { Button, Flex, Input2, Spacer } from "@artsy/palette-mobile"
 import { Expandable } from "app/Components/Expandable"
 import { dismissModal, navigate } from "app/system/navigation/navigate"
 import { useState } from "react"
@@ -11,7 +11,7 @@ export const NavigateTo: React.FC<{}> = () => {
       <Expandable label="Navigate to" expanded={false}>
         <Spacer y={1} />
         <Flex flexDirection="row">
-          <Input
+          <Input2
             placeholder="Url to navigate to"
             onChangeText={(text) => setUrl(text)}
             autoCapitalize="none"
